+++ yfinance/screener/__init__.py
@@ -1,4 +1,4 @@
-from .query import EquityQuery
-from .screener import screen, PREDEFINED_SCREENER_QUERIES
+from .query import EquityQuery, screen, FundQuery, Query, QueryHead
+from .const import PREDEFINED_SCREENER_QUERIES
 
-__all__ = ['EquityQuery', 'FundQuery', 'screen', 'PREDEFINED_SCREENER_QUERIES']
+__all__ = ["EquityQuery", "FundQuery", "screen", "PREDEFINED_SCREENER_QUERIES", "Query", "QueryHead"]
